seventh commit
the first finished functioning version of the game. Made the second enemy shoot correctly and the third to dive in the direction of the player and crash into him. made the player take damage and the game manager to track the player health and destroy him when he drops below 0